clean up basic example and create advanced example
