+++ examples/advanced/.gitignore
@@ -1,0 +1,3 @@
+.docz/
+node_modules
+dist
