+++ examples/advanced/.storybook/storybook.css
@@ -1,0 +1,3 @@
+html, body, #root {
+  height: 100%;
+}
