+++ examples/advanced/doczrc.js
@@ -1,0 +1,9 @@
+const { storybook } = require('docz-plugin-storybook')
+
+module.exports = {
+  plugins: [
+    storybook({
+      storyWrapper: require.resolve('./story-wrapper')
+    })
+  ]
+}
