+++ examples/advanced/storybook/button-story.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import Button from '../components/Button'
+import Button, { StyledButton } from '../components/Button'
 
 storiesOf('Button', module)
   .add('with text', () => <Button>Hello Button</Button>)
@@ -11,3 +11,5 @@ storiesOf('Button', module)
       </span>
     </Button>
   ))
+
+storiesOf('Styled Components', module).add('with text', () => <StyledButton>Hello Button</StyledButton>)
