+++ examples/advanced/storybook/index.js
@@ -1,0 +1,1 @@
+import './button-story'
