+++ examples/basic/doczrc.js
@@ -1,9 +1,5 @@
 const { storybook } = require('docz-plugin-storybook')
 
 module.exports = {
-  plugins: [
-    storybook({
-      storyWrapper: require.resolve('./story-wrapper')
-    })
-  ]
+  plugins: [storybook()]
 }
